Delete Temp Files on Remote PCs
Locally run script that reaches out to remote PCs to clear out defined
temporary file folders. Uses WinRM or UNC pathing. It defaults to the
WinRM method but will fail over to UNC pathing if WinRM is not enabled.
There is a commented out section that leverages PsExec and a batch file
to enable WinRM on the remote PCs. The script is heavily commented and
should answer any questions about what it does.